requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118

# --------- pytorch --------- #
torch>=2.1.0
torchvision>=0.16.0
lightning==2.1.0
torchmetrics==1.2.0

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
wandb
# neptune-client
# mlflow
comet-ml

# --------- others --------- #
pyrootutils
pre-commit
rich
pytest
sh; platform_system != 'Windows'
omegaconf
tqdm
tensorflow[and-cuda]
pynvml

# --------- data --------- #
numpy
Pillow
opencv-python-headless
albumentations
h5py

# --------- visualization --------- #
seaborn
scikit_learn
matplotlib
ipython
